Change database logic. Add DatabaseManager.
