+++ app/src/main/java/org/stepic/droid/concurrency/FromDbCoursesTask.java
@@ -8,7 +8,7 @@ import org.stepic.droid.core.IShell;
 import org.stepic.droid.events.courses.FinishingGetCoursesFromDbEvent;
 import org.stepic.droid.events.courses.StartingGetCoursesFromDbEvent;
 import org.stepic.droid.model.Course;
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 import java.util.List;
 
@@ -22,9 +22,12 @@ public class FromDbCoursesTask extends StepicTask<Void, Void, List<Course>> {
     @Inject
     Bus bus;
 
-    private DbOperationsCourses.Table mCourseType;
+    @Inject
+    DatabaseManager dbOperationsCourses;
 
-    public FromDbCoursesTask(@NotNull DbOperationsCourses.Table courseType) {
+    private DatabaseManager.Table mCourseType;
+
+    public FromDbCoursesTask(@NotNull DatabaseManager.Table courseType) {
         super(MainApplication.getAppContext());
 
         MainApplication.component(mContext).inject(this);
@@ -34,14 +37,8 @@ public class FromDbCoursesTask extends StepicTask<Void, Void, List<Course>> {
 
     @Override
     protected List<Course> doInBackgroundBody(Void... params) throws Exception {
-        DbOperationsCourses dbOperationsCourses = mShell.getDbOperationsCourses(mCourseType);
-        dbOperationsCourses.open();
         List<Course> fromCache = null;
-        try {
-            fromCache = dbOperationsCourses.getAllCourses();
-        } finally {
-            dbOperationsCourses.close();
-        }
+        fromCache = dbOperationsCourses.getAllCourses(mCourseType);
         return fromCache;
     }
