+++ app/src/main/java/org/stepic/droid/core/IShell.java
@@ -3,8 +3,6 @@ package org.stepic.droid.core;
 
 import android.content.Context;
 
-import org.stepic.droid.store.operations.DbOperationsCourses;
-import org.stepic.droid.store.operations.DbOperationsSections;
 import org.stepic.droid.preferences.SharedPreferenceHelper;
 import org.stepic.droid.web.IApi;
 
@@ -14,6 +12,4 @@ public interface IShell {
     IApi getApi();
     SharedPreferenceHelper getSharedPreferenceHelper();
     Context getContext();
-    DbOperationsCourses getDbOperationsCourses(DbOperationsCourses.Table type );
-    DbOperationsSections getDbOperationsSection();
 }
