+++ app/src/main/java/org/stepic/droid/core/Shell.java
@@ -3,8 +3,6 @@ package org.stepic.droid.core;
 import android.content.Context;
 
 import org.stepic.droid.base.MainApplication;
-import org.stepic.droid.store.operations.DbOperationsCourses;
-import org.stepic.droid.store.operations.DbOperationsSections;
 import org.stepic.droid.preferences.SharedPreferenceHelper;
 import org.stepic.droid.web.IApi;
 
@@ -50,15 +48,4 @@ public class Shell implements IShell {
     public Context getContext() {
         return mContext;
     }
-
-    @Override
-    public DbOperationsCourses getDbOperationsCourses(DbOperationsCourses.Table type) {
-        return new DbOperationsCourses(mContext, type); //how much objects here?
-    }
-
-    @Override
-    public DbOperationsSections getDbOperationsSection() {
-        return new DbOperationsSections(mContext);
-    }
-
 }
