+++ app/src/main/java/org/stepic/droid/events/courses/FailCoursesDownloadEvent.java
@@ -1,10 +1,10 @@
 package org.stepic.droid.events.courses;
 
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 public class FailCoursesDownloadEvent extends CourseEventBase {
 
-    public FailCoursesDownloadEvent(DbOperationsCourses.Table type) {
+    public FailCoursesDownloadEvent(DatabaseManager.Table type) {
         super(type);
     }
 }
