+++ app/src/main/java/org/stepic/droid/events/courses/FinishingGetCoursesFromDbEvent.java
@@ -1,14 +1,14 @@
 package org.stepic.droid.events.courses;
 
 import org.stepic.droid.model.Course;
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 import java.util.List;
 
 public class FinishingGetCoursesFromDbEvent extends CourseEventBase {
     private final List<Course> result;
 
-    public FinishingGetCoursesFromDbEvent(DbOperationsCourses.Table type, List<Course> result) {
+    public FinishingGetCoursesFromDbEvent(DatabaseManager.Table type, List<Course> result) {
         super(type);
         this.result = result;
     }
