+++ app/src/main/java/org/stepic/droid/events/courses/FinishingSaveCoursesToDbEvent.java
@@ -1,9 +1,9 @@
 package org.stepic.droid.events.courses;
 
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 public class FinishingSaveCoursesToDbEvent extends CourseEventBase {
-    public FinishingSaveCoursesToDbEvent(DbOperationsCourses.Table type) {
+    public FinishingSaveCoursesToDbEvent(DatabaseManager.Table type) {
         super(type);
     }
 }
