+++ app/src/main/java/org/stepic/droid/events/courses/GettingCoursesFromDbSuccessEvent.java
@@ -1,7 +1,7 @@
 package org.stepic.droid.events.courses;
 
 import org.stepic.droid.model.Course;
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 import java.util.List;
 
@@ -12,7 +12,7 @@ public class GettingCoursesFromDbSuccessEvent extends CourseEventBase {
 
     private final List<Course> courses;
 
-    public GettingCoursesFromDbSuccessEvent(DbOperationsCourses.Table type, List<Course> courses) {
+    public GettingCoursesFromDbSuccessEvent(DatabaseManager.Table type, List<Course> courses) {
         super(type);
         this.courses = courses;
     }
