+++ app/src/main/java/org/stepic/droid/events/courses/StartingGetCoursesFromDbEvent.java
@@ -1,9 +1,9 @@
 package org.stepic.droid.events.courses;
 
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 public class StartingGetCoursesFromDbEvent extends CourseEventBase{
-    public StartingGetCoursesFromDbEvent(DbOperationsCourses.Table type) {
+    public StartingGetCoursesFromDbEvent(DatabaseManager.Table type) {
         super(type);
     }
 }
