+++ app/src/main/java/org/stepic/droid/events/courses/StartingSaveCoursesToDbEvent.java
@@ -1,9 +1,9 @@
 package org.stepic.droid.events.courses;
 
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 public class StartingSaveCoursesToDbEvent extends CourseEventBase {
-    public StartingSaveCoursesToDbEvent(DbOperationsCourses.Table type) {
+    public StartingSaveCoursesToDbEvent(DatabaseManager.Table type) {
         super(type);
     }
 }
