+++ app/src/main/java/org/stepic/droid/store/operations/IDatabaseManager.java
@@ -1,0 +1,5 @@
+package org.stepic.droid.store.operations;
+
+public interface IDatabaseManager {
+    void dropDatabase();
+}
