--- app/src/main/java/org/stepic/droid/store/operations/IDatabaseOperations.java
@@ -1,13 +1,0 @@
-package org.stepic.droid.store.operations;
-
-import android.database.Cursor;
-
-import java.sql.SQLException;
-
-public interface IDatabaseOperations {
-    void open() throws SQLException;
-
-    void close();
-
-    Cursor getCursor();
-}
