+++ app/src/main/java/org/stepic/droid/view/fragments/MyCoursesFragment.java
@@ -13,7 +13,7 @@ import org.stepic.droid.events.courses.PreLoadCoursesEvent;
 import org.stepic.droid.events.courses.StartingGetCoursesFromDbEvent;
 import org.stepic.droid.events.courses.StartingSaveCoursesToDbEvent;
 import org.stepic.droid.events.courses.SuccessCoursesDownloadEvent;
-import org.stepic.droid.store.operations.DbOperationsCourses;
+import org.stepic.droid.store.operations.DatabaseManager;
 
 public class MyCoursesFragment extends CoursesFragmentBase {
 
@@ -32,7 +32,7 @@ public class MyCoursesFragment extends CoursesFragmentBase {
     @Override
     public void onStart() {
         super.onStart();
-        mTypeOfCourse = DbOperationsCourses.Table.enrolled;
+        mTypeOfCourse = DatabaseManager.Table.enrolled;
     }
